+++ integration_tests/suite/test_wazo_user_backend.py
@@ -19,7 +19,8 @@ class TestWazoUserBackend(WazoAuthTestCase):
                 token=uuid_(),
                 auth_id=user['uuid'],
                 xivo_user_uuid=user['uuid'],  # For API compatibility reason
-                acls=has_items('confd.#', 'plugind.#'),
+                acls=has_items('confd.#', 'plugind.#'),  # DEPRECATED
+                acl=has_items('confd.#', 'plugind.#'),
                 session_uuid=uuid_(),
                 metadata=has_entries(
                     pbx_user_uuid=user['uuid'], xivo_user_uuid=user['uuid']